feat(analyze): .data jump tables, disasm cap+pagination, audit hardening
Jump-table resolution now handles tables the compiler placed in .rdata/.data, not
just inline .text. makeJumpTableResolver reads the table from any readable section
while still requiring case targets to land in the executable section, so a switch
like D2Game 0x6fd14a5d (`jmp [eax*4+0x6fd2f8fe]`, table in .data) resolves -- its
case blocks now anchor at call-target/high instead of heuristic low.

disassemble: raise maxDisasmCount 600 -> 1000 and add a truncation hint. When the
output hits the instruction cap mid-function (not a ret/boundary) with bytes still
ahead, it prints a "** truncated ... continue with va=0x... **" note so an agent
paginates instead of assuming it saw the whole function. count default stays 50;
the handler count description now advertises the cap + stop_at_ret so agents know
to raise it (previously they had no signal the output was cut).

Audit hardening (Codex gpt-5.5 audit, findings verified against the binary):
 - jump-table case targets are filtered to in-range/unvisited before being pushed
   onto the CFG-walk stack, bounding stack growth against a crafted table whose
   entries re-point at their own switch.
 - pdata.go computes the .pdata slice end in uint64 to avoid a uint32 wrap of
   off+Size (crafted x64 PE with a huge exception-table Size) yielding end<off and
   a slice-bounds panic.

A leading-hole table finding (slot0 invalid, cases from slot1; 4 D2 sites) was
deliberately NOT adopted: skipping invalid leading slots would risk promoting
non-switch disp32 jumps to fake tables (trust > coverage); those sites fall back
safely to low.

Tests: disasm_truncate_test.go (ungated -- cap hint fires mid-function, not at a
ret boundary or on exhausted data) and TestFunctionAtSwitchCaseAnchorsViaDataTable
(gated -- the .data-table switch case).

+++ tools/analyze/pdata.go
@@ -46,11 +46,13 @@ func pdataFuncEndOffset(f *pe.File, queryRVA uint32) (uint32, bool) {
 			if uint64(off) >= uint64(len(secData)) {
 				return 0, false // .pdata offset beyond section raw data
 			}
-			end := off + excDir.Size
-			if int(end) > len(secData) {
-				end = uint32(len(secData))
+			// uint64 math avoids a uint32 wrap of off+Size (a crafted PE with a huge
+			// exception-table Size) producing end < off and a slice-bounds panic.
+			end := uint64(off) + uint64(excDir.Size)
+			if end > uint64(len(secData)) {
+				end = uint64(len(secData))
 			}
-			pdataData = secData[off:end]
+			pdataData = secData[uint64(off):end]
 			break
 		}
 	}
